ol upgrade and refactoring of map container
