+++ package.json
@@ -43,7 +43,7 @@
     "lodash": "4.17.15",
     "moment": "2.24.0",
     "ngx-color-picker": "7.3.1",
-    "ol": "5.3.0",
+    "ol": "6.1.1",
     "papaparse": "4.6.3",
     "rxjs": "6.4.0",
     "ts-helpers": "1.1.2",
