+++ src/app/map/map-container/map-container.component.scss
@@ -1,0 +1,9 @@
+.map {
+    background: black;
+    overflow: auto;
+    height: 100%;
+}
+
+mat-grid-list {
+    height: 100%;
+}
