+++ src/app/map/map.service.ts
@@ -1,16 +1,13 @@
-
 import {distinctUntilChanged} from 'rxjs/operators';
-import {Observable, BehaviorSubject, Subject} from 'rxjs';
+import {BehaviorSubject, Observable} from 'rxjs';
 
 import {Injectable} from '@angular/core';
 import {Extent as OlExtent} from 'ol'
-import {getHeight as olExtentGetHeight, getWidth as olExtentGetWidth, getIntersection as olExtentGetIntersection, containsExtent as olExtentContainsExtent} from 'ol/extent';
+import {containsExtent as olExtentContainsExtent, getIntersection as olExtentGetIntersection} from 'ol/extent';
 import {GeometryType as OlGeometryType} from 'ol/geom';
 import {Vector as OlSourceVector} from 'ol/source';
 
-import {Symbology} from '../layers/symbology/symbology.model';
-import {Layer} from '../layers/layer.model';
-import {MapComponent} from './map.component';
+import {MapContainerComponent} from './map-container/map-container.component';
 
 export interface ViewportSize {
     extent: Extent;
@@ -18,7 +15,7 @@ export interface ViewportSize {
     maxExtent?: [number, number, number, number];
 }
 
-type Extent = [number, number, number, number]  | OlExtent;
+type Extent = [number, number, number, number] | OlExtent;
 
 @Injectable()
 export class MapService {
@@ -27,37 +24,12 @@ export class MapService {
         resolution: 1,
     });
 
-    private mapComponent: MapComponent;
-    private zoomToExtent$ = new Subject<Extent>();
-    private zoomToLayer$ = new Subject<Layer<Symbology>>();
+    private mapComponent: MapContainerComponent;
 
     constructor() {
-        // this.viewportSize$.subscribe(
-        //    v => console.log('viewport', v.extent.join(','), v.resolution)
-        // );
     }
 
-    getZoomToExtentStream(): Observable<Extent> {
-        return this.zoomToExtent$;
-    }
-
-    getZoomToLayerStream(): Observable<Layer<Symbology>> {
-        return this.zoomToLayer$;
-    }
-
-    public zoomToLayer(l: Layer<Symbology>) {
-        this.zoomToLayer$.next(l);
-    }
-
-    public zoomToLayers(l: Iterable<Layer<any>>) {
-
-    }
-
-    public zoomToExtent(extent: Extent) {
-        this.zoomToExtent$.next(extent);
-    }
-
-    public registerMapComponent(mapComponent: MapComponent) {
+    public registerMapComponent(mapComponent: MapContainerComponent) {
         this.mapComponent = mapComponent;
     }
 
@@ -68,6 +40,7 @@ export class MapService {
         this.mapComponent.startDrawInteraction(drawType);
     }
 
+    // TODO: decide to use or loose it
     public isDrawInteractionAttached(): boolean {
         return this.mapComponent.isDrawInteractionAttached();
     }
@@ -86,14 +59,10 @@ export class MapService {
 
         const oldViewportSize = this.viewportSize$.value;
 
-        if (
-            this.resolutionChanged(oldViewportSize, newViewportSize)
-            || !this.extentContains(oldViewportSize, newViewportSize)
-        ) {
-
-            const w = olExtentGetWidth(newViewportSize.extent);
-            const h = olExtentGetHeight(newViewportSize.extent);
-            let newExtent = newViewportSize.extent; // ol.extent.buffer(newViewportSize.extent, Math.max(w, h) * 0.5);
+        if (resolutionChanged(oldViewportSize, newViewportSize) || !extentContains(oldViewportSize, newViewportSize)) {
+            // TODO: buffer extent to query more data
+            // ol.extent.buffer(newViewportSize.extent, Math.max(w, h) * 0.5);
+            let newExtent = newViewportSize.extent;
 
             if (newViewportSize.maxExtent) {
                 newExtent = olExtentGetIntersection(newExtent, newViewportSize.maxExtent);
@@ -112,15 +81,14 @@ export class MapService {
     getViewportSizeStream(): Observable<ViewportSize> {
         return this.viewportSize$.pipe(distinctUntilChanged());
     }
+}
 
-    private resolutionChanged(vps1: ViewportSize, vps2: ViewportSize): boolean {
-        return vps1.resolution !== vps2.resolution;
-    }
+function extentContains(vps1: ViewportSize, vps2: ViewportSize): boolean {
+    const e1 = (vps1.maxExtent) ? olExtentGetIntersection(vps1.extent, vps1.maxExtent) : vps1.extent;
+    const e2 = (vps2.maxExtent) ? olExtentGetIntersection(vps2.extent, vps2.maxExtent) : vps2.extent;
+    return olExtentContainsExtent(e1, e2);
+}
 
-    private extentContains(vps1: ViewportSize, vps2: ViewportSize): boolean {
-        const e1 = (vps1.maxExtent) ? olExtentGetIntersection(vps1.extent, vps1.maxExtent) : vps1.extent;
-        const e2 = (vps2.maxExtent) ? olExtentGetIntersection(vps2.extent, vps2.maxExtent) : vps2.extent;
-        const contains = olExtentContainsExtent(e1, e2);
-        return contains;
-    }
+function resolutionChanged(vps1: ViewportSize, vps2: ViewportSize): boolean {
+    return vps1.resolution !== vps2.resolution;
 }
