+++ tsconfig.json
@@ -1,19 +1,21 @@
 {
     "compileOnSave": false,
     "compilerOptions": {
-    "importHelpers": true,
+        "baseUrl": "./",
         "outDir": "./dist/out-tsc",
         "sourceMap": true,
         "declaration": false,
+        "module": "es2015",
         "moduleResolution": "node",
         "emitDecoratorMetadata": true,
         "experimentalDecorators": true,
+        "importHelpers": true,
         "target": "es5",
         "typeRoots": [
             "node_modules/@types"
         ],
         "lib": [
-            "es2016",
+            "es2018",
             "dom"
         ]
     },
